updated all compile/target versions; updated to androidx test classes; removed legacy non-partitioned tests
