+++ azuredata/src/androidTest/java/com/azure/data/integration/DocumentManualPartitionKeyTests.kt
@@ -1,6 +1,6 @@
 package com.azure.data.integration
 
-import android.support.test.runner.AndroidJUnit4
+import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.azure.data.AzureData
 import com.azure.data.integration.common.CustomDocument
 import com.azure.data.model.service.Response
