+++ azuredata/src/androidTest/java/com/azure/data/integration/ResourceOracleTests.kt
@@ -1,6 +1,6 @@
 package com.azure.data.integration
 
-import android.support.test.runner.AndroidJUnit4
+import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.azure.data.model.DocumentCollection
 import com.azure.data.model.Resource
 import com.azure.core.util.ContextProvider
