+++ build.gradle
@@ -2,7 +2,7 @@
 
 buildscript {
 
-    ext.azureandroid_release_version = '0.3.0'
+    ext.azureandroid_release_version = '0.4.0'
     ext.kotlin_version = '1.3.50'
     ext.gradle_version = '3.5.0'
 
@@ -33,10 +33,11 @@ allprojects {
         okhttpVersion = "4.1.1"
         gsonVersion = "2.8.5"
         jUnitVersion = "4.12"
-        espressoVersion = "3.0.1"
         reactiveNetworkVersion = "0.12.3"
-        testRunnerVersion = "1.0.2"
-        testRulesVersion = "1.0.2"
+        espressoVersion = "3.1.0"
+        testRunnerVersion = "1.2.0"
+        testRulesVersion = "1.2.0"
+        jUnitRunnerVersion = "1.1.1"
         timberVersion = "4.7.1"
         awaitilityVersion = "3.0.0"
     }
